First Commit
This is the first commit with the structure of the project. It is ready
to work. The connection to MongoDB is ready just need to change the
mongoldb connection data to the local instance. I created the User
entity, the UserRepository and its test as a reference for the other
entities.
The next commit will come with a RESTful service as a reference.