Quick fix: Remove users issue
The fix was that it was not possible to remove users from the database. Now the users
can be removed as long as the user does not have any assignment assigned at the moment
when it is tried to be removed. Otherwise, the server will response with an error
saying that the user still has assignments assigned.